// schemas.d.ts
export interface ApiEventResponseEventResponse extends CollectionTypeSchema {
  info: {
    singularName: 'event-response';
    pluralName: 'event-responses';
    displayName: 'Event Response';
    description: '';
  };
  options: {
    draftAndPublish: false;
  };
  attributes: {
    name: StringAttribute & RequiredAttribute;
    email: EmailAttribute & RequiredAttribute;
    restriction_vegetarian: BooleanAttribute &
      RequiredAttribute &
      DefaultTo<false>;
    restriction_vegan: BooleanAttribute & RequiredAttribute & DefaultTo<false>;
    restriction_nuts: BooleanAttribute & RequiredAttribute & DefaultTo<false>;
    restriction_dairy: BooleanAttribute & RequiredAttribute & DefaultTo<false>;
    restriction_alcohol: BooleanAttribute &
      RequiredAttribute &
      DefaultTo<false>;
    restriction_other: StringAttribute;
    comment: StringAttribute;
    termsOfService: BooleanAttribute & RequiredAttribute;
    event: RelationAttribute<
      'api::event-response.event-response',
      'manyToOne',
      'api::event.event'
    >;
    confirm_token: StringAttribute & UniqueAttribute;
    state: EnumerationAttribute<
      ['waiting', 'invite_sent', 'invite_confirmed', 'invite_cancelled']
    > &
      RequiredAttribute &
      DefaultTo<'waiting'>;
    invite_date: DateTimeAttribute;
    createdAt: DateTimeAttribute;
    updatedAt: DateTimeAttribute;
    createdBy: RelationAttribute<
      'api::event-response.event-response',
      'oneToOne',
      'admin::user'
    > &
      PrivateAttribute;
    updatedBy: RelationAttribute<
      'api::event-response.event-response',
      'oneToOne',
      'admin::user'
    > &
      PrivateAttribute;
  };
}